+++ sdk/tests/core/test_az_json.c
@@ -57,7 +57,7 @@ static void test_json_reader_init(void** state)
   assert_int_equal(az_json_reader_init(&reader, AZ_SPAN_FROM_STR("\""), NULL), AZ_OK);
   assert_int_equal(az_json_reader_init(&reader, AZ_SPAN_FROM_STR("\""), &options), AZ_OK);
 
-  test_json_token_helper(reader.token, AZ_JSON_TOKEN_NONE, AZ_SPAN_NULL);
+  test_json_token_helper(reader.token, AZ_JSON_TOKEN_NONE, AZ_SPAN_EMPTY);
 }
 
 /**  Json writer **/
@@ -223,7 +223,7 @@ static void test_json_writer(void** state)
   }
   {
     az_json_writer writer = { 0 };
-    TEST_EXPECT_SUCCESS(az_json_writer_init(&writer, AZ_SPAN_NULL, NULL));
+    TEST_EXPECT_SUCCESS(az_json_writer_init(&writer, AZ_SPAN_EMPTY, NULL));
     assert_int_equal(az_json_writer_append_int32(&writer, 1), AZ_ERROR_INSUFFICIENT_SPAN_SIZE);
   }
 }
@@ -408,7 +408,7 @@ az_result test_allocator_always_null(
     az_span* out_next_destination)
 {
   (void)allocator_context;
-  *out_next_destination = AZ_SPAN_NULL;
+  *out_next_destination = AZ_SPAN_EMPTY;
 
   return AZ_OK;
 }
@@ -423,7 +423,7 @@ static void test_json_writer_chunked(void** state)
     _az_user_context user_context = { .current_index = &previous };
 
     TEST_EXPECT_SUCCESS(
-        az_json_writer_chunked_init(&writer, AZ_SPAN_NULL, allocator, (void*)&user_context, NULL));
+        az_json_writer_chunked_init(&writer, AZ_SPAN_EMPTY, allocator, (void*)&user_context, NULL));
 
     // 0___________________________________________________________________________________________________1
     // 0_________1_________2_________3_________4_________5_________6_________7_________8_________9_________0
@@ -492,7 +492,7 @@ static void test_json_writer_chunked(void** state)
 
     {
       TEST_EXPECT_SUCCESS(az_json_writer_chunked_init(
-          &writer, AZ_SPAN_NULL, allocator, (void*)&user_context, NULL));
+          &writer, AZ_SPAN_EMPTY, allocator, (void*)&user_context, NULL));
 
       TEST_EXPECT_SUCCESS(az_json_writer_append_double(&writer, 0.000000000000001, 15));
 
@@ -510,7 +510,7 @@ static void test_json_writer_chunked(void** state)
     }
     {
       TEST_EXPECT_SUCCESS(az_json_writer_chunked_init(
-          &writer, AZ_SPAN_NULL, allocator, (void*)&user_context, NULL));
+          &writer, AZ_SPAN_EMPTY, allocator, (void*)&user_context, NULL));
 
       TEST_EXPECT_SUCCESS(az_json_writer_append_double(&writer, 1e-300, 15));
 
@@ -535,7 +535,7 @@ static void test_json_writer_chunked(void** state)
     _az_user_context user_context = { .current_index = &previous };
 
     TEST_EXPECT_SUCCESS(
-        az_json_writer_chunked_init(&writer, AZ_SPAN_NULL, allocator, (void*)&user_context, NULL));
+        az_json_writer_chunked_init(&writer, AZ_SPAN_EMPTY, allocator, (void*)&user_context, NULL));
 
     // this json { "span": "\" } would be scaped to { "span": "\\"" }
     uint8_t single_char[1] = { '\\' }; // char = '\'
@@ -572,7 +572,7 @@ static void test_json_writer_chunked(void** state)
     _az_user_context user_context = { .current_index = &previous };
 
     TEST_EXPECT_SUCCESS(
-        az_json_writer_chunked_init(&writer, AZ_SPAN_NULL, allocator, (void*)&user_context, NULL));
+        az_json_writer_chunked_init(&writer, AZ_SPAN_EMPTY, allocator, (void*)&user_context, NULL));
 
     // this json { "array": [1, 2, {}, 3, -12.3 ] }
     TEST_EXPECT_SUCCESS(az_json_writer_append_begin_object(&writer));
@@ -616,7 +616,7 @@ static void test_json_writer_chunked(void** state)
     _az_user_context user_context = { .current_index = &previous };
 
     TEST_EXPECT_SUCCESS(az_json_writer_chunked_init(
-        &nested_object_builder, AZ_SPAN_NULL, allocator, (void*)&user_context, NULL));
+        &nested_object_builder, AZ_SPAN_EMPTY, allocator, (void*)&user_context, NULL));
 
     {
       // 0___________________________________________________________________________________________________1
@@ -654,7 +654,7 @@ static void test_json_writer_chunked(void** state)
     az_json_writer writer = { 0 };
     az_span_allocator_fn allocator = &test_allocator_always_null;
 
-    TEST_EXPECT_SUCCESS(az_json_writer_chunked_init(&writer, AZ_SPAN_NULL, allocator, NULL, NULL));
+    TEST_EXPECT_SUCCESS(az_json_writer_chunked_init(&writer, AZ_SPAN_EMPTY, allocator, NULL, NULL));
     assert_int_equal(az_json_writer_append_int32(&writer, 1), AZ_ERROR_INSUFFICIENT_SPAN_SIZE);
   }
 }
@@ -900,7 +900,7 @@ static void test_json_writer_large_string_chunked(void** state)
     _az_user_context user_context = { .current_index = &previous };
 
     TEST_EXPECT_SUCCESS(
-        az_json_writer_chunked_init(&writer, AZ_SPAN_NULL, allocator, (void*)&user_context, NULL));
+        az_json_writer_chunked_init(&writer, AZ_SPAN_EMPTY, allocator, (void*)&user_context, NULL));
 
     TEST_EXPECT_SUCCESS(az_json_writer_append_string(
         &writer, az_span_slice(AZ_SPAN_FROM_BUFFER(expected), 1, 1299)));
@@ -970,7 +970,7 @@ static void test_json_reader(void** state)
     az_json_reader reader = { 0 };
     TEST_EXPECT_SUCCESS(az_json_reader_init(&reader, AZ_SPAN_FROM_STR("    "), NULL));
     assert_true(az_json_reader_next_token(&reader) == AZ_ERROR_UNEXPECTED_END);
-    test_json_token_helper(reader.token, AZ_JSON_TOKEN_NONE, AZ_SPAN_NULL);
+    test_json_token_helper(reader.token, AZ_JSON_TOKEN_NONE, AZ_SPAN_EMPTY);
   }
   {
     az_json_reader reader = { 0 };
@@ -982,7 +982,7 @@ static void test_json_reader(void** state)
     az_json_reader reader = { 0 };
     TEST_EXPECT_SUCCESS(az_json_reader_init(&reader, AZ_SPAN_FROM_STR("  nul"), NULL));
     assert_true(az_json_reader_next_token(&reader) == AZ_ERROR_UNEXPECTED_END);
-    test_json_token_helper(reader.token, AZ_JSON_TOKEN_NONE, AZ_SPAN_NULL);
+    test_json_token_helper(reader.token, AZ_JSON_TOKEN_NONE, AZ_SPAN_EMPTY);
   }
   {
     az_json_reader reader = { 0 };
@@ -994,7 +994,7 @@ static void test_json_reader(void** state)
     az_json_reader reader = { 0 };
     TEST_EXPECT_SUCCESS(az_json_reader_init(&reader, AZ_SPAN_FROM_STR("  falsx  "), NULL));
     assert_true(az_json_reader_next_token(&reader) == AZ_ERROR_UNEXPECTED_CHAR);
-    test_json_token_helper(reader.token, AZ_JSON_TOKEN_NONE, AZ_SPAN_NULL);
+    test_json_token_helper(reader.token, AZ_JSON_TOKEN_NONE, AZ_SPAN_EMPTY);
   }
   {
     az_json_reader reader = { 0 };
@@ -1012,7 +1012,7 @@ static void test_json_reader(void** state)
     az_json_reader reader = { 0 };
     TEST_EXPECT_SUCCESS(az_json_reader_init(&reader, AZ_SPAN_FROM_STR("  123a"), NULL));
     assert_true(az_json_reader_next_token(&reader) == AZ_ERROR_UNEXPECTED_CHAR);
-    test_json_token_helper(reader.token, AZ_JSON_TOKEN_NONE, AZ_SPAN_NULL);
+    test_json_token_helper(reader.token, AZ_JSON_TOKEN_NONE, AZ_SPAN_EMPTY);
   }
   {
     az_span const s = AZ_SPAN_FROM_STR(" \"tr\\\"ue\\t\" ");
@@ -1788,7 +1788,7 @@ static void test_json_value(void** state)
 #define _az_json_token_is_text_equal_hello_helper(token) \
   do \
   { \
-    assert_true(az_json_token_is_text_equal(&json_string, AZ_SPAN_NULL) == false); \
+    assert_true(az_json_token_is_text_equal(&json_string, AZ_SPAN_EMPTY) == false); \
     assert_true(az_json_token_is_text_equal(&json_string, AZ_SPAN_FROM_STR("")) == false); \
     assert_true(az_json_token_is_text_equal(&json_string, AZ_SPAN_FROM_STR("a")) == false); \
     assert_true(az_json_token_is_text_equal(&json_string, AZ_SPAN_FROM_STR("hello")) == false); \
@@ -1811,7 +1811,7 @@ static void test_json_value(void** state)
 #define _az_json_token_is_text_equal_name_helper(token) \
   do \
   { \
-    assert_true(az_json_token_is_text_equal(&json_string, AZ_SPAN_NULL) == false); \
+    assert_true(az_json_token_is_text_equal(&json_string, AZ_SPAN_EMPTY) == false); \
     assert_true(az_json_token_is_text_equal(&json_string, AZ_SPAN_FROM_STR("")) == false); \
     assert_true(az_json_token_is_text_equal(&json_string, AZ_SPAN_FROM_STR("a")) == false); \
     assert_true( \
@@ -1852,7 +1852,7 @@ static void test_az_json_token_get_string_and_text_equal(void** state)
   assert_true(az_json_token_is_text_equal(&json_string, AZ_SPAN_FROM_STR("a")) == false);
   assert_true(az_json_token_is_text_equal(&json_string, AZ_SPAN_FROM_STR("aaaaaa")) == false);
   assert_true(az_json_token_is_text_equal(&json_string, AZ_SPAN_FROM_STR("       ")) == false);
-  assert_true(az_json_token_is_text_equal(&json_string, AZ_SPAN_NULL));
+  assert_true(az_json_token_is_text_equal(&json_string, AZ_SPAN_EMPTY));
   assert_true(az_json_token_is_text_equal(&json_string, AZ_SPAN_FROM_STR("")));
 
   json_string.kind = AZ_JSON_TOKEN_PROPERTY_NAME;
@@ -1862,7 +1862,7 @@ static void test_az_json_token_get_string_and_text_equal(void** state)
   assert_true(az_json_token_is_text_equal(&json_string, AZ_SPAN_FROM_STR("a")) == false);
   assert_true(az_json_token_is_text_equal(&json_string, AZ_SPAN_FROM_STR("aaaaaa")) == false);
   assert_true(az_json_token_is_text_equal(&json_string, AZ_SPAN_FROM_STR("       ")) == false);
-  assert_true(az_json_token_is_text_equal(&json_string, AZ_SPAN_NULL));
+  assert_true(az_json_token_is_text_equal(&json_string, AZ_SPAN_EMPTY));
   assert_true(az_json_token_is_text_equal(&json_string, AZ_SPAN_FROM_STR("")));
 
   json_string = (az_json_token){
@@ -2574,7 +2574,7 @@ static void test_az_json_token_copy(void** state)
     az_span_copy(buffers[1], az_span_slice_to_end(json, 7));
 
     json_string = (az_json_token){
-      .slice = AZ_SPAN_NULL,
+      .slice = AZ_SPAN_EMPTY,
       .size = 12,
       ._internal = {
         .pointer_to_first_buffer = buffers,
@@ -2602,7 +2602,7 @@ static void test_az_json_token_copy(void** state)
     az_span_copy(buffers[2], az_span_slice_to_end(json, 9));
 
     json_string = (az_json_token){
-      .slice = AZ_SPAN_NULL,
+      .slice = AZ_SPAN_EMPTY,
       .size = 12,
       ._internal = {
         .pointer_to_first_buffer = buffers,
@@ -2623,7 +2623,7 @@ static void test_az_json_token_copy(void** state)
 
     json_string = (az_json_token){
       .kind = AZ_JSON_TOKEN_STRING,
-      .slice = AZ_SPAN_NULL,
+      .slice = AZ_SPAN_EMPTY,
       .size = 12,
       ._internal = {
         .string_has_escaped_chars = false,
@@ -2712,7 +2712,7 @@ static void test_az_json_reader_chunked(void** state)
   model original = (model){
     .name_string = property_value,
     .name_length = 32,
-    .name_value_span = AZ_SPAN_NULL,
+    .name_value_span = AZ_SPAN_EMPTY,
     .code = 0,
   };